+++ sdk/appconfiguration/Azure.Data.AppConfiguration/src/ConfigurationClient.cs
@@ -169,7 +169,7 @@ namespace Azure.Data.AppConfiguration
             BuildUriForKvRoute(request.Uri, setting);
 
             MatchConditions requestOptions = new MatchConditions();
-            requestOptions.SetIfNotExistsCondition();
+            requestOptions.IfNoneMatch = ETag.All;
             ConditionalRequestOptionsExtensions.ApplyHeaders(request, requestOptions);
 
             request.Headers.Add(s_mediaTypeKeyValueApplicationHeader);
@@ -220,7 +220,7 @@ namespace Azure.Data.AppConfiguration
             if (onlyIfUnchanged)
             {
                 requestOptions = new MatchConditions();
-                requestOptions.SetIfUnmodifiedCondition(setting.ETag);
+                requestOptions.IfMatch = setting.ETag;
             }
 
             return await SetAsync(setting, requestOptions, cancellationToken).ConfigureAwait(false);
@@ -241,7 +241,7 @@ namespace Azure.Data.AppConfiguration
             if (onlyIfUnchanged)
             {
                 requestOptions = new MatchConditions();
-                requestOptions.SetIfUnmodifiedCondition(setting.ETag);
+                requestOptions.IfMatch = setting.ETag;
             }
 
             return Set(setting, requestOptions, cancellationToken);
@@ -373,7 +373,7 @@ namespace Azure.Data.AppConfiguration
             if (onlyIfUnchanged)
             {
                 requestOptions = new MatchConditions();
-                requestOptions.SetIfUnmodifiedCondition(setting.ETag);
+                requestOptions.IfMatch = setting.ETag;
             }
 
             return await DeleteAsync(setting.Key, setting.Label, requestOptions, cancellationToken).ConfigureAwait(false);
@@ -394,7 +394,7 @@ namespace Azure.Data.AppConfiguration
             if (onlyIfUnchanged)
             {
                 requestOptions = new MatchConditions();
-                requestOptions.SetIfUnmodifiedCondition(setting.ETag);
+                requestOptions.IfMatch = setting.ETag;
             }
 
             return Delete(setting.Key, setting.Label, requestOptions, cancellationToken);
@@ -524,7 +524,7 @@ namespace Azure.Data.AppConfiguration
             if (onlyIfChanged)
             {
                 requestOptions = new MatchConditions();
-                requestOptions.SetIfModifiedCondition(setting.ETag);
+                requestOptions.IfNoneMatch = setting.ETag;
             }
 
             return await GetAsync(setting.Key, setting.Label, acceptDateTime, requestOptions, cancellationToken).ConfigureAwait(false);
@@ -546,7 +546,7 @@ namespace Azure.Data.AppConfiguration
             if (onlyIfChanged)
             {
                 requestOptions = new MatchConditions();
-                requestOptions.SetIfModifiedCondition(setting.ETag);
+                requestOptions.IfNoneMatch = setting.ETag;
             }
 
             return Get(setting.Key, setting.Label, acceptDateTime, requestOptions, cancellationToken);
@@ -705,7 +705,7 @@ namespace Azure.Data.AppConfiguration
                     case 200:
                     case 206:
                         SettingBatch settingBatch = await ConfigurationServiceSerializer.ParseBatchAsync(response, cancellationToken).ConfigureAwait(false);
-                        return new Page<ConfigurationSetting>(settingBatch.Settings, settingBatch.NextBatchLink, response);
+                        return Page<ConfigurationSetting>.FromValues(settingBatch.Settings, settingBatch.NextBatchLink, response);
                     default:
                         throw await response.CreateRequestFailedExceptionAsync().ConfigureAwait(false);
                 }
@@ -738,7 +738,7 @@ namespace Azure.Data.AppConfiguration
                     case 200:
                     case 206:
                         SettingBatch settingBatch = ConfigurationServiceSerializer.ParseBatch(response);
-                        return new Page<ConfigurationSetting>(settingBatch.Settings, settingBatch.NextBatchLink, response);
+                        return Page<ConfigurationSetting>.FromValues(settingBatch.Settings, settingBatch.NextBatchLink, response);
                     default:
                         throw response.CreateRequestFailedException();
                 }
@@ -786,7 +786,7 @@ namespace Azure.Data.AppConfiguration
                     case 200:
                     case 206:
                         SettingBatch settingBatch = await ConfigurationServiceSerializer.ParseBatchAsync(response, cancellationToken).ConfigureAwait(false);
-                        return new Page<ConfigurationSetting>(settingBatch.Settings, settingBatch.NextBatchLink, response);
+                        return Page<ConfigurationSetting>.FromValues(settingBatch.Settings, settingBatch.NextBatchLink, response);
                     default:
                         throw await response.CreateRequestFailedExceptionAsync().ConfigureAwait(false);
                 }
@@ -819,7 +819,7 @@ namespace Azure.Data.AppConfiguration
                     case 200:
                     case 206:
                         SettingBatch settingBatch = ConfigurationServiceSerializer.ParseBatch(response);
-                        return new Page<ConfigurationSetting>(settingBatch.Settings, settingBatch.NextBatchLink, response);
+                        return Page<ConfigurationSetting>.FromValues(settingBatch.Settings, settingBatch.NextBatchLink, response);
                     default:
                         throw response.CreateRequestFailedException();
                 }