+++ sdk/core/Azure.Core/src/MatchConditions.cs
@@ -17,42 +17,5 @@ namespace Azure
         /// Optionally limit requests to resources that do not match the ETag.
         /// </summary>
         public ETag? IfNoneMatch { get; set; }
-
-        /// <summary>
-        /// Set preconditions that indicate to apply an operation only if the
-        /// target resource has changed.
-        /// </summary>
-        /// <param name="etag">The version of the resource to compare to.</param>
-        public virtual void SetIfModifiedCondition(ETag etag)
-        {
-            IfNoneMatch = etag;
-        }
-
-        /// <summary>
-        /// Set preconditions that indicate to apply an operation only if the
-        /// target resource has not changed.
-        /// </summary>
-        public virtual void SetIfUnmodifiedCondition(ETag etag)
-        {
-            IfMatch = etag;
-        }
-
-        /// <summary>
-        /// Set preconditions that indicate to apply an operation only if the
-        /// target resource is present.
-        /// </summary>
-        public virtual void SetIfExistsCondition()
-        {
-            IfMatch = ETag.All;
-        }
-
-        /// <summary>
-        /// Set preconditions that indicate to apply an operation only if the
-        /// target resource is not present.
-        /// </summary>
-        public virtual void SetIfNotExistsCondition()
-        {
-            IfNoneMatch = ETag.All;
-        }
     }
 }