+++ sdk/storage/Azure.Storage.Blobs/tests/ContainerClientTests.cs
@@ -10,8 +10,6 @@ using System.Threading.Tasks;
 using Azure.Core.Testing;
 using Azure.Storage.Blobs.Models;
 using Azure.Storage.Blobs.Specialized;
-using Azure.Storage.Common;
-using Azure.Storage.Common.Test;
 using Azure.Storage.Test;
 using Azure.Storage.Test.Shared;
 using NUnit.Framework;
@@ -586,9 +584,9 @@ namespace Azure.Storage.Blobs.Test
 
                 SignedIdentifier acl = response.Value.SignedIdentifiers.First();
                 Assert.AreEqual(signedIdentifiers[0].Id, acl.Id);
-                Assert.AreEqual(signedIdentifiers[0].AccessPolicy.Start, acl.AccessPolicy.Start);
-                Assert.AreEqual(signedIdentifiers[0].AccessPolicy.Expiry, acl.AccessPolicy.Expiry);
-                Assert.AreEqual(signedIdentifiers[0].AccessPolicy.Permission, acl.AccessPolicy.Permission);
+                Assert.AreEqual(signedIdentifiers[0].AccessPolicy.StartsOn, acl.AccessPolicy.StartsOn);
+                Assert.AreEqual(signedIdentifiers[0].AccessPolicy.ExpiresOn, acl.AccessPolicy.ExpiresOn);
+                Assert.AreEqual(signedIdentifiers[0].AccessPolicy.Permissions, acl.AccessPolicy.Permissions);
             }
         }
 
@@ -682,7 +680,7 @@ namespace Azure.Storage.Blobs.Test
             var duration = TimeSpan.FromSeconds(15);
 
             // Act
-            Response<Lease> response = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration: duration);
+            Response<BlobLease> response = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration: duration);
 
             // Assert
             Assert.AreEqual(id, response.Value.LeaseId);
@@ -745,7 +743,7 @@ namespace Azure.Storage.Blobs.Test
                 var duration = TimeSpan.FromSeconds(15);
 
                 // Act
-                Response<Lease> response = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(
+                Response<BlobLease> response = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(
                     duration: duration,
                     accessConditions: accessConditions.HttpAccessConditions);
 
@@ -800,11 +798,11 @@ namespace Azure.Storage.Blobs.Test
             var id = Recording.Random.NewGuid().ToString();
             var duration = TimeSpan.FromSeconds(15);
 
-            Response<Lease> leaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(
+            Response<BlobLease> leaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(
                 duration: duration);
 
             // Act
-            Response<Lease> renewResponse = await InstrumentClient(container.GetLeaseClient(leaseResponse.Value.LeaseId)).RenewAsync();
+            Response<BlobLease> renewResponse = await InstrumentClient(container.GetLeaseClient(leaseResponse.Value.LeaseId)).RenewAsync();
 
             // Assert
             Assert.IsNotNull(renewResponse.GetRawResponse().Headers.RequestId);
@@ -853,7 +851,7 @@ namespace Azure.Storage.Blobs.Test
                 _ = await lease.AcquireAsync(duration: duration);
 
                 // Act
-                Response<Lease> response = await lease.RenewAsync(accessConditions: accessConditions.HttpAccessConditions);
+                Response<BlobLease> response = await lease.RenewAsync(accessConditions: accessConditions.HttpAccessConditions);
 
                 // Assert
                 Assert.IsNotNull(response.GetRawResponse().Headers.RequestId);
@@ -887,7 +885,7 @@ namespace Azure.Storage.Blobs.Test
                 var duration = TimeSpan.FromSeconds(15);
 
                 LeaseClient lease = InstrumentClient(container.GetLeaseClient(id));
-                Response<Lease> aquireLeaseResponse = await lease.AcquireAsync(duration: duration);
+                Response<BlobLease> aquireLeaseResponse = await lease.AcquireAsync(duration: duration);
 
                 // Act
                 await TestHelper.AssertExpectedExceptionAsync<StorageRequestFailedException>(
@@ -913,7 +911,7 @@ namespace Azure.Storage.Blobs.Test
                 // Arrange
                 var id = Recording.Random.NewGuid().ToString();
                 var duration = TimeSpan.FromSeconds(15);
-                Response<Lease> leaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration);
+                Response<BlobLease> leaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration);
 
                 // Act
                 Response<ReleasedObjectInfo> releaseResponse = await InstrumentClient(container.GetLeaseClient(leaseResponse.Value.LeaseId)).ReleaseAsync();
@@ -957,7 +955,7 @@ namespace Azure.Storage.Blobs.Test
                     var duration = TimeSpan.FromSeconds(15);
 
                     LeaseClient lease = InstrumentClient(container.GetLeaseClient(id));
-                    Response<Lease> aquireLeaseResponse = await lease.AcquireAsync(duration: duration);
+                    Response<BlobLease> aquireLeaseResponse = await lease.AcquireAsync(duration: duration);
 
                     // Act
                     Response<ReleasedObjectInfo> response = await lease.ReleaseAsync(accessConditions: accessConditions.HttpAccessConditions);
@@ -986,7 +984,7 @@ namespace Azure.Storage.Blobs.Test
                 var duration = TimeSpan.FromSeconds(15);
 
                 LeaseClient lease = InstrumentClient(container.GetLeaseClient(id));
-                Response<Lease> aquireLeaseResponse = await lease.AcquireAsync(duration: duration);
+                Response<BlobLease> aquireLeaseResponse = await lease.AcquireAsync(duration: duration);
 
                 // Act
                 await TestHelper.AssertExpectedExceptionAsync<StorageRequestFailedException>(
@@ -1016,7 +1014,7 @@ namespace Azure.Storage.Blobs.Test
                 TimeSpan breakPeriod = TimeSpan.FromSeconds(0);
 
                 // Act
-                Response<Lease> breakResponse = await InstrumentClient(container.GetLeaseClient()).BreakAsync(breakPeriod);
+                Response<BlobLease> breakResponse = await InstrumentClient(container.GetLeaseClient()).BreakAsync(breakPeriod);
 
                 // Assert
                 Response<BlobContainerItem> response = await container.GetPropertiesAsync();
@@ -1057,10 +1055,10 @@ namespace Azure.Storage.Blobs.Test
                 var id = Recording.Random.NewGuid().ToString();
                 var duration = TimeSpan.FromSeconds(15);
 
-                Response<Lease> aquireLeaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration: duration);
+                Response<BlobLease> aquireLeaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration: duration);
 
                 // Act
-                Response<Lease> response = await InstrumentClient(container.GetLeaseClient()).BreakAsync(
+                Response<BlobLease> response = await InstrumentClient(container.GetLeaseClient()).BreakAsync(
                     accessConditions: accessConditions.HttpAccessConditions);
 
                 // Assert
@@ -1094,7 +1092,7 @@ namespace Azure.Storage.Blobs.Test
                 var id = Recording.Random.NewGuid().ToString();
                 var duration = TimeSpan.FromSeconds(15);
 
-                Response<Lease> aquireLeaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration: duration);
+                Response<BlobLease> aquireLeaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration: duration);
 
                 // Act
                 await TestHelper.AssertExpectedExceptionAsync<StorageRequestFailedException>(
@@ -1121,11 +1119,11 @@ namespace Azure.Storage.Blobs.Test
                 // Arrange
                 var id = Recording.Random.NewGuid().ToString();
                 var duration = TimeSpan.FromSeconds(15);
-                Response<Lease> leaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration);
+                Response<BlobLease> leaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration);
                 var newId = Recording.Random.NewGuid().ToString();
 
                 // Act
-                Response<Lease> changeResponse = await InstrumentClient(container.GetLeaseClient(id)).ChangeAsync(newId);
+                Response<BlobLease> changeResponse = await InstrumentClient(container.GetLeaseClient(id)).ChangeAsync(newId);
 
                 // Assert
                 Assert.AreEqual(newId, changeResponse.Value.LeaseId);
@@ -1168,10 +1166,10 @@ namespace Azure.Storage.Blobs.Test
                 var newId = Recording.Random.NewGuid().ToString();
                 var duration = TimeSpan.FromSeconds(15);
 
-                Response<Lease> aquireLeaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration: duration);
+                Response<BlobLease> aquireLeaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration: duration);
 
                 // Act
-                Response<Lease> response = await InstrumentClient(container.GetLeaseClient(aquireLeaseResponse.Value.LeaseId)).ChangeAsync(
+                Response<BlobLease> response = await InstrumentClient(container.GetLeaseClient(aquireLeaseResponse.Value.LeaseId)).ChangeAsync(
                     proposedId: newId,
                     accessConditions: accessConditions.HttpAccessConditions);
 
@@ -1207,7 +1205,7 @@ namespace Azure.Storage.Blobs.Test
                 var newId = Recording.Random.NewGuid().ToString();
                 var duration = TimeSpan.FromSeconds(15);
 
-                Response<Lease> aquireLeaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration: duration);
+                Response<BlobLease> aquireLeaseResponse = await InstrumentClient(container.GetLeaseClient(id)).AcquireAsync(duration: duration);
 
                 // Act
                 await TestHelper.AssertExpectedExceptionAsync<StorageRequestFailedException>(
@@ -1391,7 +1389,6 @@ namespace Azure.Storage.Blobs.Test
                 e => Assert.AreEqual("ContainerNotFound", e.ErrorCode.Split('\n')[0]));
         }
 
-        [Ignore("https://github.com/Azure/azure-sdk-for-net/issues/8113")]
         [Test]
         public async Task ListBlobsFlatSegmentAsync_PreservesWhitespace()
         {
@@ -1607,7 +1604,7 @@ namespace Azure.Storage.Blobs.Test
                 using var stream = new MemoryStream(GetRandomBuffer(100));
                 await container.UploadBlobAsync(name, stream);
                 Response<BlobProperties> properties = await InstrumentClient(container.GetBlobClient(name)).GetPropertiesAsync();
-                Assert.AreEqual(BlobType.BlockBlob, properties.Value.BlobType);
+                Assert.AreEqual(BlobType.Block, properties.Value.BlobType);
             }
         }