+++ sdk/storage/Azure.Storage.Common/src/PartitionedDownloader.cs
@@ -7,7 +7,6 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Azure.Storage.Common;
 
 namespace Azure.Storage
 {
@@ -127,7 +126,7 @@ namespace Azure.Storage
                     parallelTransferOptions ??= new ParallelTransferOptions();
 
                     var maximumThreadCount =
-                        parallelTransferOptions.Value.MaximumThreadCount ?? Constants.Blob.Block.DefaultConcurrentTransfersCount;
+                        parallelTransferOptions.Value.MaximumConcurrency ?? Constants.Blob.Block.DefaultConcurrentTransfersCount;
                     var maximumPartitionLength =
                         Math.Min(
                             Constants.Blob.Block.MaxDownloadBytes,