+++ sdk/storage/Azure.Storage.Common/src/StorageResponseClassifier.cs
@@ -5,7 +5,7 @@ using System;
 using Azure.Core;
 using Azure.Core.Pipeline;
 
-namespace Azure.Storage.Common
+namespace Azure.Storage
 {
     internal class StorageResponseClassifier : ResponseClassifier
     {