+++ sdk/storage/Azure.Storage.Common/tests/Shared/TestExceptionPolicy.cs
@@ -8,7 +8,7 @@ using Azure.Core;
 using Azure.Core.Pipeline;
 using Azure.Core.Testing;
 
-namespace Azure.Storage.Common.Test
+namespace Azure.Storage.Test
 {
     public class TestExceptionPolicy : HttpPipelinePolicy
     {