+++ sdk/storage/Azure.Storage.Files/src/Models/StorageFileDownloadInfo.cs
@@ -31,6 +31,11 @@ namespace Azure.Storage.Files.Models
         public Stream Content => _flattened.Content;
 
         /// <summary>
+        /// The content type specified for the file. The default content type is 'application/octet-stream'
+        /// </summary>
+        public string ContentType => _flattened.ContentType;
+
+        /// <summary>
         /// If the file has an MD5 hash and this operation is to read the full content, this response header is returned so that the client can check for message content integrity.
         /// </summary>
 #pragma warning disable CA1819 // Properties should not return arrays