+++ sdk/storage/Azure.Storage.Queues/src/QueueClient.cs
@@ -9,7 +9,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using Azure.Core;
 using Azure.Core.Pipeline;
-using Azure.Storage.Common;
 using Azure.Storage.Queues.Models;
 using Metadata = System.Collections.Generic.IDictionary<string, string>;