+++ sdk/storage/Microsoft.Azure.Management.Storage/changelog.md
@@ -1,5 +1,11 @@
 ## Microsoft.Azure.Management.Storage release notes
 
+### Changes in 13.3.0
+
+- Support enable Files Azure Active Directory Domain Service Authentication when create or update Storage account
+- Support regenerateKey for Kerberos keys on Storage account
+- Support list Kerberos keys on Storage account
+
 ### Changes in 13.2.0
 
 - Support set LargeFileSharesState as Enabled on Create or Update Storage account